backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
python-dotenv
pytesseract
pdf2image
pillow
sentence-transformers
faiss-cpu
spacy
presidio-analyzer
jinja2
pdfkit
requests
cryptography
PyPDF2
reportlab
rq
rq-dashboard
gunicorn
